requirements.txt
jieba==0.42.1
jsonlines==4.0.0
ollama==0.6.1
pydantic==2.12.4
PyYAML==6.0.3
rank_bm25==0.2.2
tqdm==4.67.1
nltk==3.9.1
scikit-learn==1.5.2
faiss-cpu
numpy
openai
pysbd
rouge
rouge-chinese
sentence-transformers
transformers
langchain
langchain-community
langchain-text-splitters
langchain-core
requests
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama